roast/requirements.txt
opencv-python
ultralytics
requests
python-dotenv
